+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-## Next
+## 0.0.50
 ### Features
+* New `<range-and-bearing>` component that is used to to draw range & bearing on the map.
+* Refactored `KeyboardControlService` docs.
+* Added `maximumNumberOfPoints` to `PolygonEditOptions` - you can now set maximum number of points when creating polygon using `PolygonEditorService` .
+* Refactored `KeyboardControlService` docs. 
 * Added flyTo options to `CameraService.trackEntity()` + example in the demo.
+
 ### Fixes
 * `ac-point-desc` changed to Entity instead of Primitive. For primitive use `ac-point-primitive-desc`.